Move toggle method to navbar component
