+++ src/App.js
@@ -1,4 +1,3 @@
-import React, { useState } from 'react';
 import { Routes, Route } from 'react-router-dom';
 import Splash from './components/pages/splash';
 import Login from './components/pages/login';
@@ -6,24 +5,10 @@ import Signup from './components/pages/signup';
 import Navbar from './layout/Navbar';
 
 function App() {
-  const [classValue, setClassValue] = useState('hide');
-  // const [showNav, setShowNav] = useState('block');
-  const toggleMenu = () => {
-    if (classValue === 'show') {
-      setClassValue('hide');
-      // setShowNav('none');
-    } else {
-      setClassValue('show');
-      // setShowNav('block');
-    }
-  };
   return (
     <div className="main-container">
       <div>
-        <Navbar
-          classValue={classValue}
-          toggleMenu={toggleMenu}
-        />
+        <Navbar />
       </div>
       <div className="container-fluid px-0">
         <Routes>
